overlay: elvees: systemd: Add CAN configuration
In BusyBox the inital interfaces configuration was stored
in /etc/network/interfaces. This includes the configuration
for CAN.

With the switch to systemd the configuration from
/etc/network/interfaces must be ported as .network files. There is
no such file for CAN by default. The CAN interface requires an
inital baud rate to be set in order to function.

This commit adds a configuration file which sets initial baud rate
for CAN.

Issue: #MCOM02SW-916
Change-Id: Id89871a1b5932b2f17f6c775b7166e257792bf22